pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lifeguard"
version = "0.1.0"
description = "Lifeguard Discord bot"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "discord.py>=2.4.0",
  "python-dotenv>=1.0.1",
  "firebase-admin>=6.5.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

